requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dateutil
starlette
